New: Refactor UiModel for transactions.
Utilize a StateFlow, instead, and also move the UiModel logic from the fragment into the viewModel.
