+++ app/src/main/java/cash/z/ecc/android/ui/history/TransactionFragment.kt
@@ -19,14 +19,15 @@ import cash.z.ecc.android.ext.*
 import cash.z.ecc.android.feedback.Report
 import cash.z.ecc.android.sdk.db.entity.ConfirmedTransaction
 import cash.z.ecc.android.sdk.ext.ZcashSdk
-import cash.z.ecc.android.sdk.ext.convertZatoshiToZecString
 import cash.z.ecc.android.sdk.ext.toAbbreviatedAddress
 import cash.z.ecc.android.sdk.ext.twig
 import cash.z.ecc.android.ui.MainActivity
 import cash.z.ecc.android.ui.base.BaseFragment
+import cash.z.ecc.android.ui.history.HistoryViewModel.UiModel
 import cash.z.ecc.android.ui.util.toUtf8Memo
+import kotlinx.coroutines.flow.collect
+import kotlinx.coroutines.flow.stateIn
 import kotlinx.coroutines.launch
-import java.text.NumberFormat
 import java.util.*
 
 
@@ -71,12 +72,12 @@ class TransactionFragment : BaseFragment<FragmentTransactionBinding>() {
         super.onViewCreated(view, savedInstanceState)
 
         binding.apply {
-            ViewCompat.setTransitionName(topBoxValue, "test_amount_anim_${viewModel.selectedTransaction!!.id}")
-            ViewCompat.setTransitionName(topBoxBackground, "test_bg_anim_${viewModel.selectedTransaction!!.id}")
+            ViewCompat.setTransitionName(topBoxValue, "test_amount_anim_${viewModel.selectedTransaction.value?.id}")
+            ViewCompat.setTransitionName(topBoxBackground, "test_bg_anim_${viewModel.selectedTransaction.value?.id}")
             backButtonHitArea.onClickNavBack { tapped(Report.Tap.TRANSACTION_BACK) }
 
             lifecycleScope.launch {
-                viewModel.selectedTransaction.toUiModel(viewModel.latestHeight).let { uiModel ->
+                viewModel.uiModels.stateIn(lifecycleScope).collect { uiModel ->
                     topBoxLabel.text = uiModel.topLabel
                     topBoxValue.text = uiModel.topValue
                     bottomBoxLabel.text = uiModel.bottomLabel
@@ -190,112 +191,9 @@ class TransactionFragment : BaseFragment<FragmentTransactionBinding>() {
         }
     }
 
-    private suspend fun ConfirmedTransaction?.toUiModel(latestHeight: Int? = null): UiModel = UiModel().apply {
-        this@toUiModel.let { tx ->
-            txId = mainActivity?.toTxId(tx?.rawTransactionId)
-            isInbound = when {
-                !(tx?.toAddress.isNullOrEmpty()) -> false
-                tx != null && tx.toAddress.isNullOrEmpty() && tx.value > 0L && tx.minedHeight > 0 -> true
-                else -> null
-            }
-            isMined = tx?.minedHeight != null && tx.minedHeight > ZcashSdk.SAPLING_ACTIVATION_HEIGHT
-            topValue = if (tx == null) "" else "\$${WalletZecFormmatter.toZecStringFull(tx.value)}"
-            minedHeight = (tx?.minedHeight ?: 0).toString()
-            val flags =
-                DateUtils.FORMAT_SHOW_DATE or DateUtils.FORMAT_SHOW_YEAR or DateUtils.FORMAT_ABBREV_MONTH
-            timestamp = if (tx == null) getString(R.string.transaction_timestamp_unavailable) else DateUtils.getRelativeDateTimeString(
-                ZcashWalletApp.instance,
-                tx.blockTimeInSeconds * 1000,
-                DateUtils.SECOND_IN_MILLIS,
-                DateUtils.WEEK_IN_MILLIS,
-                flags
-            ).toString()
 
-            // memo logic
-            val txMemo = tx?.memo.toUtf8Memo()
-            if (!txMemo.isEmpty()) {
-                memo = txMemo
-            }
 
-            // confirmation logic
-            // TODO: clean all of this up and remove/improve reliance on `isSufficientlyOld` function. Also, add a constant for the number of confirmations we expect.
-            tx?.let {
-                val isMined = it.blockTimeInSeconds != 0L
-                if (isMined) {
-                    val hasLatestHeight = latestHeight != null && latestHeight > ZcashSdk.SAPLING_ACTIVATION_HEIGHT
-                    if (it.minedHeight > 0 && hasLatestHeight) {
-                        val confirmations = latestHeight!! - it.minedHeight + 1
-                        confirmation = if (confirmations >= 10) getString(R.string.transaction_status_confirmed) else "$confirmations ${getString(R.string.transaction_status_confirming)}"
-                    } else {
-                        if (!hasLatestHeight && isSufficientlyOld(tx)) {
-                            twig("Warning: could not load latestheight from server to determine confirmations but this transaction is mined and old enough to be considered confirmed")
-                            confirmation = getString(R.string.transaction_status_confirmed)
-                        } else {
-                            twig("Warning: could not determine confirmation text value so it will be left null!")
-                            confirmation = getString(R.string.transaction_confirmation_count_unavailable)
-                        }
-                    }
-                } else {
-                    confirmation = getString(R.string.transaction_status_pending)
-                }
 
-            }
-
-            val mainActivity = (context as MainActivity)
-            // inbound v. outbound values
-            when (isInbound) {
-                true -> {
-                    topLabel = getString(R.string.transaction_story_inbound)
-                    bottomLabel = getString(R.string.transaction_story_inbound_total)
-                    bottomValue = "\$${WalletZecFormmatter.toZecStringFull(tx?.value)}"
-                    iconRotation = 315f
-                    source = getString(R.string.transaction_story_to_shielded)
-                    address = mainActivity.extractValidAddress(tx?.memo.toUtf8Memo())
-                }
-                false -> {
-                    topLabel = getString(R.string.transaction_story_outbound)
-                    bottomLabel = getString(R.string.transaction_story_outbound_total)
-                    bottomValue = "\$${WalletZecFormmatter.toZecStringFull(tx?.value?.plus(ZcashSdk.MINERS_FEE_ZATOSHI))}"
-                    iconRotation = 135f
-                    fee = getString(R.string.transaction_story_network_fee, WalletZecFormmatter.toZecStringFull(ZcashSdk.MINERS_FEE_ZATOSHI))
-                    source = getString(R.string.transaction_story_from_shielded)
-                    address = tx?.toAddress
-                }
-                null -> {
-                    twig("Error: transaction appears to be invalid.")
-                }
-            }
-        }
-    }
-
-    // TODO: determine this in a more generic and technically correct way. For now, this is good enough.
-    //       the goal is just to improve the edge cases where the latest height isn't known but other
-    //       information suggests that the TX is confirmed. We can improve this, later.
-    private fun isSufficientlyOld(tx: ConfirmedTransaction): Boolean {
-        val threshold = 75 * 1000 * 25 //approx 25 blocks
-        val delta = System.currentTimeMillis() / 1000L - tx.blockTimeInSeconds
-        return tx.minedHeight > ZcashSdk.SAPLING_ACTIVATION_HEIGHT
-                && delta < threshold
-    }
-
-    data class UiModel(
-        var topLabel: String = "",
-        var topValue: String = "",
-        var bottomLabel: String = "",
-        var bottomValue: String = "",
-        var minedHeight: String = "",
-        var timestamp: String = "",
-        var iconRotation: Float = -1f,
-
-        var fee: String? = null,
-        var source: String? = null,
-        var memo: String? = null,
-        var address: String? = null,
-        var isInbound: Boolean? = null,
-        var isMined: Boolean = false,
-        var confirmation: String? = null,
-        var txId: String? = null
-    )
 }
